+++ autd3capi-link-visualizer/src/lib.rs
@@ -77,13 +77,13 @@ macro_rules! match_visualizer_plot {
 #[no_mangle]
 #[must_use]
 pub unsafe extern "C" fn AUTDLinkVisualizerPlotRange(
-    x_min: float,
-    x_max: float,
-    y_min: float,
-    y_max: float,
-    z_min: float,
-    z_max: float,
-    resolution: float,
+    x_min: f64,
+    x_max: f64,
+    y_min: f64,
+    y_max: f64,
+    z_min: f64,
+    z_max: f64,
+    resolution: f64,
 ) -> PlotRangePtr {
     PlotRangePtr(Box::into_raw(Box::new(PlotRange {
         x_range: x_min..x_max,
@@ -96,7 +96,7 @@ pub unsafe extern "C" fn AUTDLinkVisualizerPlotRange(
 #[no_mangle]
 #[must_use]
 pub unsafe extern "C" fn AUTDLinkVisualizerPlotRangeObservePointsLen(range: PlotRangePtr) -> u64 {
-    let n = |range: &std::ops::Range<float>, resolution: float| -> usize {
+    let n = |range: &std::ops::Range<f64>, resolution: f64| -> usize {
         ((range.end - range.start) / resolution).floor() as usize + 1
     };
     let nx = n(&range.x_range, range.resolution);
@@ -108,12 +108,12 @@ pub unsafe extern "C" fn AUTDLinkVisualizerPlotRangeObservePointsLen(range: Plot
 #[no_mangle]
 pub unsafe extern "C" fn AUTDLinkVisualizerPlotRangeObservePoints(
     range: PlotRangePtr,
-    points: *mut float,
+    points: *mut f64,
 ) {
     let range = take!(range, PlotRange);
     let observe_points = range.observe_points();
     std::ptr::copy_nonoverlapping(
-        observe_points.as_ptr() as *const float,
+        observe_points.as_ptr() as *const f64,
         points,
         observe_points.len() * 3,
     );
@@ -197,12 +197,12 @@ pub unsafe extern "C" fn AUTDLinkVisualizerCalcField(
     visualizer: LinkPtr,
     backend: Backend,
     directivity: Directivity,
-    points: *const float,
+    points: *const f64,
     points_len: u32,
     geometry: GeometryPtr,
     segment: Segment,
     idx: u32,
-    buf: *mut float,
+    buf: *mut f64,
 ) -> ResultI32 {
     let idx = idx as usize;
     let segment = segment.into();
