+++ autd3capi/src/geometry/device.rs
@@ -19,39 +19,39 @@ pub unsafe extern "C" fn AUTDDeviceNumTransducers(dev: DevicePtr) -> u32 {
 
 #[no_mangle]
 #[must_use]
-pub unsafe extern "C" fn AUTDDeviceGetSoundSpeed(dev: DevicePtr) -> float {
+pub unsafe extern "C" fn AUTDDeviceGetSoundSpeed(dev: DevicePtr) -> f64 {
     dev.sound_speed
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn AUTDDeviceSetSoundSpeed(mut dev: DevicePtr, value: float) {
+pub unsafe extern "C" fn AUTDDeviceSetSoundSpeed(mut dev: DevicePtr, value: f64) {
     dev.sound_speed = value;
 }
 
 #[no_mangle]
 pub unsafe extern "C" fn AUTDDeviceSetSoundSpeedFromTemp(
     mut dev: DevicePtr,
-    temp: float,
-    k: float,
-    r: float,
-    m: float,
+    temp: f64,
+    k: f64,
+    r: f64,
+    m: f64,
 ) {
     dev.set_sound_speed_from_temp_with(temp, k, r, m);
 }
 
 #[no_mangle]
 #[must_use]
-pub unsafe extern "C" fn AUTDDeviceGetAttenuation(dev: DevicePtr) -> float {
+pub unsafe extern "C" fn AUTDDeviceGetAttenuation(dev: DevicePtr) -> f64 {
     dev.attenuation
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn AUTDDeviceSetAttenuation(mut dev: DevicePtr, value: float) {
+pub unsafe extern "C" fn AUTDDeviceSetAttenuation(mut dev: DevicePtr, value: f64) {
     dev.attenuation = value;
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn AUTDDeviceCenter(dev: DevicePtr, center: *mut float) {
+pub unsafe extern "C" fn AUTDDeviceCenter(dev: DevicePtr, center: *mut f64) {
     let c = dev.center();
     center.add(0).write(c.x);
     center.add(1).write(c.y);
@@ -59,31 +59,25 @@ pub unsafe extern "C" fn AUTDDeviceCenter(dev: DevicePtr, center: *mut float) {
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn AUTDDeviceTranslate(mut dev: DevicePtr, x: float, y: float, z: float) {
+pub unsafe extern "C" fn AUTDDeviceTranslate(mut dev: DevicePtr, x: f64, y: f64, z: f64) {
     dev.translate(Vector3::new(x, y, z));
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn AUTDDeviceRotate(
-    mut dev: DevicePtr,
-    w: float,
-    i: float,
-    j: float,
-    k: float,
-) {
+pub unsafe extern "C" fn AUTDDeviceRotate(mut dev: DevicePtr, w: f64, i: f64, j: f64, k: f64) {
     dev.rotate(UnitQuaternion::from_quaternion(Quaternion::new(w, i, j, k)));
 }
 
 #[no_mangle]
 pub unsafe extern "C" fn AUTDDeviceAffine(
     mut dev: DevicePtr,
-    x: float,
-    y: float,
-    z: float,
-    w: float,
-    i: float,
-    j: float,
-    k: float,
+    x: f64,
+    y: f64,
+    z: f64,
+    w: f64,
+    i: f64,
+    j: f64,
+    k: f64,
 ) {
     dev.affine(
         Vector3::new(x, y, z),
